fix stuff and init stuff

+++ include/chisel/chassis/drive.h
@@ -21,12 +21,12 @@ struct DriveTrain {
     );
 };
 
-constexpr auto LINEAR_CURVE = [](const float input) {return input;};
-constexpr auto SIGMOID_CURVE = [](const float input) {
+constexpr auto LINEAR_CURVE = [](const int32_t input) {return input;};
+constexpr auto SIGMOID_CURVE = [](const int32_t input) {
     // finer control at high speeds
     return 127 * std::sin((M_PI * input) / 254);
 };
-constexpr auto SIGMA_CURVE = [](const float input) {
+constexpr auto SIGMA_CURVE = [](const int32_t input) {
     // finer control at low speeds
     return 254 * 1 / M_PI / std::asin(input / 127);
 };
